+++ ais/emd_internal_test.go
@@ -5,7 +5,6 @@
 package ais
 
 import (
-	"strconv"
 	"testing"
 	"time"
 
@@ -19,25 +18,6 @@ import (
 	. "github.com/onsi/ginkgo/v2"
 	. "github.com/onsi/gomega"
 )
-
-const mockPodSpec = `
-apiVersion: v1
-kind: Pod
-metadata:
-  name: mock-pod-spec
-spec:
-  containers:
-  - name: main
-    image: busybox
-    command: ["sleep", "3600"]
-    ports:
-    - name: default
-      containerPort: 80
-    readinessProbe:
-      httpGet:
-        path: /health
-        port: default
-`
 
 func TestEtlMDDeepCopy(t *testing.T) {
 	etlMD := newEtlMD()
@@ -103,35 +83,16 @@ var _ = Describe("EtlMD marshal and unmarshal", func() {
 		cfg = cmn.GCO.Get()
 
 		etlMD = newEtlMD()
-		for _, initType := range []string{etl.ETLSpecType, etl.SpecType} {
-			for i := range 1 {
-				var msg etl.InitMsg
-				if initType == etl.ETLSpecType {
-					msg = &etl.ETLSpecMsg{
-						InitMsgBase: etl.InitMsgBase{
-							EtlName:     "runtime-spec" + strconv.Itoa(i),
-							CommTypeX:   etl.Hpush,
-							InitTimeout: cos.Duration(etl.DefaultInitTimeout),
-							ObjTimeout:  cos.Duration(etl.DefaultObjTimeout),
-						},
-						Runtime: etl.RuntimeSpec{
-							Image: "test-runtime-image",
-						},
-					}
-				} else {
-					msg = &etl.InitSpecMsg{
-						InitMsgBase: etl.InitMsgBase{
-							EtlName:     "init-spec" + strconv.Itoa(i),
-							CommTypeX:   etl.Hpush,
-							InitTimeout: cos.Duration(etl.DefaultInitTimeout),
-							ObjTimeout:  cos.Duration(etl.DefaultObjTimeout),
-						},
-						Spec: []byte(mockPodSpec),
-					}
-				}
-				etlMD.Add(msg, etl.Running, make(etl.PodMap, 4)) // Running stage expects no-nil pod map
-			}
+		msg := &etl.ETLSpecMsg{
+			InitMsgBase: etl.InitMsgBase{
+				EtlName:     "runtime-spec",
+				CommTypeX:   etl.Hpush,
+				InitTimeout: cos.Duration(etl.DefaultInitTimeout),
+				ObjTimeout:  cos.Duration(etl.DefaultObjTimeout),
+			},
+			Runtime: etl.RuntimeSpec{Image: "test-runtime-image"},
 		}
+		etlMD.Add(msg, etl.Running, make(etl.PodMap, 4)) // Running stage expects no-nil pod map
 	})
 
 	for _, node := range []string{apc.Target, apc.Proxy} {